tcdb_api_tdd/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-limiter>=3.3.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
gudhi>=3.5.0
joblib>=1.0.0
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-flask>=1.2.0
requests>=2.28.0
pyjwt>=2.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0
